fix naming, add tests
